test: unify account provider mocks
