+++ packages/fx-core/tests/component/driver/apiKey/update.test.ts
@@ -16,12 +16,8 @@ import {
   ApiSecretRegistrationAppType,
   ApiSecretRegistrationTargetAudience,
 } from "../../../../src/component/driver/teamsApp/interfaces/ApiSecretRegistration";
-import {
-  MockedAzureAccountProvider,
-  MockedLogProvider,
-  MockedM365Provider,
-  MockedUserInteraction,
-} from "../../../plugins/solution/util";
+import { MockedLogProvider, MockedUserInteraction } from "../../../plugins/solution/util";
+import { MockedAzureAccountProvider, MockedM365Provider } from "../../../core/utils";
 
 chai.use(chaiAsPromised);
 const expect = chai.expect;
