+++ packages/fx-core/tests/component/driver/script/dotnetBuildDriver.test.ts
@@ -7,11 +7,10 @@ import * as sinon from "sinon";
 import * as tools from "../../../../src/common/utils";
 import * as utils from "../../../../src/component/driver/script/scriptDriver";
 import { DotnetBuildDriver } from "../../../../src/component/driver/script/dotnetBuildDriver";
-import { TestAzureAccountProvider } from "../../util/azureAccountMock";
 import { TestLogProvider } from "../../util/logProviderMock";
 import { DriverContext } from "../../../../src/component/driver/interface/commonArgs";
 import { assert } from "chai";
-import { MockUserInteraction } from "../../../core/utils";
+import { MockedAzureAccountProvider, MockUserInteraction } from "../../../core/utils";
 import { err, IProgressHandler, ok, UserError } from "@microsoft/teamsfx-api";
 
 describe("Dotnet Build Driver test", () => {
@@ -39,7 +38,7 @@ describe("Dotnet Build Driver test", () => {
       execPath: "/usr/local/bin",
     };
     const context = {
-      azureAccountProvider: new TestAzureAccountProvider(),
+      azureAccountProvider: new MockedAzureAccountProvider(),
       ui: new MockUserInteraction(),
       logProvider: new TestLogProvider(),
       projectPath: "./",
@@ -58,7 +57,7 @@ describe("Dotnet Build Driver test", () => {
       args: "build",
     };
     const context = {
-      azureAccountProvider: new TestAzureAccountProvider(),
+      azureAccountProvider: new MockedAzureAccountProvider(),
       ui: new MockUserInteraction(),
       logProvider: new TestLogProvider(),
       projectPath: "./",
@@ -76,7 +75,7 @@ describe("Dotnet Build Driver test", () => {
       args: "build",
     };
     const context = {
-      azureAccountProvider: new TestAzureAccountProvider(),
+      azureAccountProvider: new MockedAzureAccountProvider(),
       logProvider: new TestLogProvider(),
       projectPath: "./",
     } as any;
