+++ packages/fx-core/tests/component/feature/collaboration.test.ts
@@ -6,15 +6,12 @@ import * as sinon from "sinon";
 import * as chai from "chai";
 import chaiAsPromised from "chai-as-promised";
 import { AadCollaboration, TeamsCollaboration } from "../../../src/component/feature/collaboration";
-import {
-  MockedLogProvider,
-  MockedM365Provider,
-  MockedV2Context,
-} from "../../plugins/solution/util";
+import { MockedLogProvider, MockedV2Context } from "../../plugins/solution/util";
 import { AadAppClient } from "../../../src/component/driver/aad/utility/aadAppClient";
 import axios from "axios";
 import { AppUser } from "../../../src/component/driver/teamsApp/interfaces/appdefinitions/appUser";
 import { teamsDevPortalClient } from "../../../src/client/teamsDevPortalClient";
+import { MockedM365Provider } from "../../core/utils";
 
 chai.use(chaiAsPromised);
 const expect = chai.expect;
